Fix code smell
