+++ src/VirtoCommerce.Platform.Web/Startup.cs
@@ -508,42 +508,9 @@ namespace VirtoCommerce.Platform.Web
                 RequestPath = new PathString($"/$(Platform)/Scripts")
             });
 
-            var localModules = app.ApplicationServices.GetRequiredService<ILocalModuleCatalog>().Modules;
-            foreach (var module in localModules.OfType<ManifestModuleInfo>())
-            {
-                // Step 1. Enables static file serving for module
-                app.UseStaticFiles(new StaticFileOptions()
-                {
-                    FileProvider = new PhysicalFileProvider(module.FullPhysicalPath),
-                    RequestPath = new PathString($"/modules/$({module.ModuleName})")
-                });
 
-                // Step 2. Enables static file serving for apps
-                foreach (var moduleApp in module.Apps)
-                {
-                    var appPath = string.IsNullOrEmpty(moduleApp.ContentPath) ?
-                        Path.Combine(module.FullPhysicalPath, "Content", moduleApp.Id):
-                        Path.Combine(module.FullPhysicalPath, moduleApp.ContentPath);
-
-                    if (!Directory.Exists(appPath))
-                    {
-                        throw new ModuleInitializeException($"The '{appPath}' directory doesn't exist for {module.ModuleName}");
-                    }
-
-                    app.UseDefaultFiles(new DefaultFilesOptions()
-                    {
-                        FileProvider = new PhysicalFileProvider(appPath),
-                        RequestPath = new PathString($"/apps/{moduleApp.Id}")
-                    });
-                    app.UseStaticFiles(new StaticFileOptions()
-                    {
-                        FileProvider = new PhysicalFileProvider(appPath),
-                        RequestPath = new PathString($"/apps/{moduleApp.Id}")
-                    });
-                }
-            }
-
-
+            // Enables static file serving with the module and apps options
+            app.UseModulesAndAppsFiles();
 
             app.UseDefaultFiles();
